[v0.90.1][runtime] Continue splitting long_lived_agent.rs responsibilities (Closes #2208)
